requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# CORS and API
django-cors-headers==4.3.1
django-filter==23.3

# Image processing
Pillow==10.1.0

# Environment and configuration
python-decouple==3.8

# Database (use psycopg2-binary for easier installation)
# If this fails, try: pip install psycopg2-binary --no-binary :all:
psycopg2-binary==2.9.9

# Production server
gunicorn==21.2.0

# Static files
whitenoise==6.6.0

# Development tools (optional)
# django-debug-toolbar==4.2.0
